password reset
user is able to reset the password now. Created a component that handles a request password reset and a component that accepts a new passwored a token that is passed through a query. I'm using nodemailer to send the reset e-mail to the user. Also slight styling, removed console logs