changed the style of pagination, removed unused files
decided to change the pagination from displaying pages to loading more on click. I'm using the fetchMore() function that apollo provides for this but it has required some tweaks in apollo inMemoryCache in order to work. The incoming data has to be merged with the cached existing data. I've removed the unused files and components and did some styling adjustments