added addToCart mutation, styling
I added a mutation that will control adding a new item to the cart. The mutation checks if the cart already has an item in it (existingCartItem), if so the quantity will be incremented by one. If not a new cartItem will be created. Each cartItem has a relationship with the user and a product