update product, useForm update
Added the option for the user to update the product information. I'm using a mutation to control which information gets updated. Also i'm passing my product query based on the id of the product that the user selects. The values of the product are being passed to the useForm function as an initialState. Because it is an object i'm using Object.values to grab the specific values. useEffect is necessary as out initialState is empty until the data comes from the server, so the initialState needs to be put to the useEffect first and our values come as a dependency array