fixed pagination, refactoring
Manged to get the pagination working. Realized that i could just query the user and the necessary fields without querying authenticatedItem. Had to remove orders and products fields from the authenticatedItem query because it was clashing with other queries with the same name. Also realized that i could merge cache of the specific field in typePolicies