fixed deleteProduct mutation, readme
I had to update the deleteProduct mutation as it was not clearing the cache on succesfull delete. I added cache.gc() as per apollos docs, it will remove all the cache that is not reachable any more. Also had to update Products fetch policy as the cache is being evicted now, so i'm checking both network and cache